# chore: Use faster JSON encoding library