`scip print --json` is a bit slow; this change provides a noticeable speedup.

E.g. on benchmarking against [flink-1.scip](https://drive.google.com/drive/folders/1z62Se7eHaa5T89a16-y7s0Z1qbRY4VCg).

```
❯ hyperfine '~/Code/scip/scip-before print --json flink-1.scip' '~/Code/scip/scip-after print --json flink-1.scip'
Benchmark 1: ~/Code/scip/scip-before print --json flink-1.scip
  Time (mean ± σ):      5.786 s ±  0.547 s    [User: 7.225 s, System: 0.861 s]
  Range (min … max):    5.499 s …  7.330 s    10 runs

Benchmark 2: ~/Code/scip/scip-after print --json flink-1.scip
  Time (mean ± σ):      3.323 s ±  0.197 s    [User: 4.294 s, System: 0.685 s]
  Range (min … max):    3.180 s …  3.870 s    10 runs

Summary
  '~/Code/scip/scip-after print --json flink-1.scip' ran
    1.74 ± 0.19 times faster than '~/Code/scip/scip-before print --json flink-1.scip'
```

### Test plan

n/a